Update main.c
Network interface was not being init. This was due to the network being init AFTER the call to OSStart(). OSStart only returns if a critical error was encountered. 

This commit now inits the network interface before calling OSStart. This makes sure the network is up and running before RTOS is started. Commit also adds checks to every network function to stop on error. If an error is encountered then a debugger can be used to stop execution and check the error code.